move pgpass functions into seperate submodule
